Introduce 'o' option for accurate backup
This feature can be used -only- if all of the daemons (DIR, SD and FD)
are updated to this version.

When the new 'o' accurate option is used, only metadata is being backed up in case file's
contents did not change (meaning that file's checksum stays the same).
New stream was added to handle that during the backup and restore
phase.