Fix #7628 About 'reload' command crashing the director
This issue was not seen when trying to start the director with the same,
broken config file. It's because we use different error codes for config
parsing when the dir is being started and when we do the reload. For the
startup, we use M_ERROR_TERM which in the end just stops the director.
In case of reload we use M_ERROR (since we don't want the dir to stop in
case of trying to load broken config), so the error from parsing config
was not propagated anywhere (err msg was printed, but the parser's
caller could be unaware about it). New field was added to inform about
the error, besides printing err message.