metaplugin: Add a proper job cancel handling.
With this patch you can receive information about a cancel event
as a SIGUSR1 signal. Now the backend can intercept this signal to
handle clean and finish procedures. After that for about 5 secs,
which is the default but you can customize it using
CUSTOMCANCELSLEEP compile-in variable, the backend receive
a termination - bpipe communication will be closed and SIGTERM
sent. It is assumed that after receiving SIGUSR1 no standard
metaplugin protocol communication is expected and supported.
In this case the metaplugin return bRC_Error on every plugin
entrypoint except freePlugin().