Fix org#2623 About .ls/estimate command not printing files correctly
In my Messages resource in FD config I don't have 'Restored' flag set:

Messages {
  Name = "Standard"
  Director = darkstar-dir = Info, Warning, Error, Fatal, Terminate, Saved, Security, Alert
}

In bacula/src/lib/attr.c:259 in print_ls_output() we have the following condition:

   /* No need to compute everything if it's not required */
   if (!chk_dbglvl(dbglvl) && !is_message_type_set(jcr, message_type)) {
      return;
   }

The dbglvl is set to 150 and message_type default is set to M_RESTORED. With my
FD configuration this condition was true but when I set debug level > 150 the
condition was false and .ls listing started working.

The workaround is to set Restored flag in the FD config in the Director
messages resource.