metaplugin: Add accurate mode query.
This patch implements a basic accurate mode query to the metaplugin
protocol. In this extension the backend can issue `CHECK:$fname`
command followed by a standard `STAT:...` and `TSTAMP:...` commands
to ask Bacula if it should backup $fname in this job.
The metaplugin will response with simple "OK", you should backup this
file or "SEEN" it before, backup is not required. This feature will work
with Accurate jobs only. If job is not accurate then user will get
an error job message.